[Refactor] #183 - UIImage Type 자체에 비즈니스 로직이 담겨있던 구조 개선
